Adds a new primary row designation for trajectories based on location. Also updates some comments
There was an issue where trajectories in rows with a large y-dimension
deviation were not able to be sorted into rows based on their respective
nearest neighbor. This update adds a simpler sorting method that relies
on comparing the y distance between a new traj and the previous traj to
the distance from one traj to its nearest neighbor. If the y distance is
smaller, the row count stays the same.